# app/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "dbadeeds.ai REST API",
    "description": "AI Database Intelligence Platform API — fully replaces Streamlit session_state with stateless REST endpoints.",
    "version": "3.1.0",
    "contact": { "name": "Your Name", "email": "your-email@example.com" }
  },
  "servers": [{ "url": "/api", "description": "Flask/Gunicorn backend" }],
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "CookieAuth": { "type": "apiKey", "in": "cookie", "name": "access_token_cookie" }
    }
  },
  "security": [{ "BearerAuth": [] }, { "CookieAuth": [] }],
  "paths": {
...
    "/users/{username}": { "get":{"tags":["Users"],"summary":"Get user","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"User"}}},"put":{"tags":["Users"],"summary":"Update user","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Updated"}}},"delete":{"tags":["Users"],"summary":"Delete user","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Deleted"}}}},
...
  }
}